delete user

+++ frontend/src/app/admin/users/page.js
@@ -94,14 +94,23 @@ const Users = () => {
         setIsPreviewModalOpen(true);
     };
 
-    const handleDeleteClick = async (id) => {
-        try {
-            await axios.delete(`http://localhost:3001/api/users/${id}`);
-            fetchUsers(); 
-        } catch (error) {
-            console.error('Error deleting user:', error);
+    const handleDeleteClick = (id) => {
+        // Show confirmation dialog
+        const confirmDelete = window.confirm("Are you sure you want to delete this user?");
+        
+        // If user confirms deletion, proceed with delete operation
+        if (confirmDelete) {
+            axios.delete(`http://localhost:3001/api/users/${id}`)
+                .then(() => {
+                    fetchUsers(); // Fetch updated user list
+                })
+                .catch(error => {
+                    console.error('Error deleting user:', error);
+                });
         }
     };
+    
+
     
     
     
@@ -273,6 +282,7 @@ const Users = () => {
                                                     )}
                                                 </td>
                                             </tr>
+                                            
                                         ))}
                                     </tbody>
                                 </table>
@@ -450,9 +460,10 @@ const Users = () => {
 </div>
 
 {/* <!-- Delete modal --> */}
+{/*
 <div id="deleteModal" tabindex="-1" aria-hidden="true" class="hidden overflow-y-auto overflow-x-hidden fixed top-0 right-0 left-0 z-50 justify-center items-center w-full md:inset-0 h-[calc(100%-1rem)] max-h-full">
     <div class="relative p-4 w-full max-w-md max-h-full">
-        {/* <!-- Modal content --> */}
+        {/* <!-- Modal content --> */}{/*
         <div class="relative p-4 text-center bg-white rounded-lg shadow dark:bg-gray-800 sm:p-5">
             <button type="button" class="text-gray-400 absolute top-2.5 right-2.5 bg-transparent hover:bg-gray-200 hover:text-gray-900 rounded-lg text-sm p-1.5 ml-auto inline-flex items-center dark:hover:bg-gray-600 dark:hover:text-white" data-modal-toggle="deleteModal">
                 <svg aria-hidden="true" class="w-5 h-5" fill="currentColor" viewbox="0 0 20 20" xmlns="http://www.w3.org/2000/svg">
@@ -470,7 +481,7 @@ const Users = () => {
             </div>
         </div>
     </div>
-</div>
+            </div>*/}
 </div> 
   )
 }
